Authorize all http requests for endpoints, Handle new user registration, Handle top 10 user overall ranking, Nickname and coins are displayed on home page.

+++ backend/src/main/java/com/example/demoio/models/DisplayRankingData.java
@@ -1,4 +1,4 @@
 package com.example.demoio.models;
 
-public record DisplayRankingData(String username, int score, int coins) {
+public record DisplayRankingData(String username, double score, double coins) {
 }
